MT-29 move icon files
