+++ src/components/Header/index.tsx
@@ -1,6 +1,6 @@
 import { Container, ContentsWrapper, Contents } from './styles';
 import { useSticky } from '@hooks/useSticky';
-import { UserIcon } from '@assets/svg/UserIcon';
+import { UserIcon } from '@icons/UserIcon';
 import { theme }from '@styles/theme'
 import Logo from '@components/atom/Logo';
 import { useIsLocationHome } from '@/hooks/useIsLocationHome';
