+++ src/icons/ArrowUpIcon.tsx
