+++ src/icons/SocialGoogleIconFill.tsx
